Remove crypting2
crypting.py same as crypting2.py
