+++ crypting.py
@@ -8,49 +8,49 @@ Created on Wed Aug  5 13:04:21 2020
 from cryptography.fernet import Fernet
 
 
-def write_key():
+def write_key(loc):
     """
     Generates a key and save it into a file
     """
     key = Fernet.generate_key()
-    with open("key.key", "wb") as key_file:
+    with open(loc+"key.key", "wb") as key_file:
         key_file.write(key)
 
 
-def load_key():
+def load_key(loc):
     """
     Loads the key from the current directory named `key.key`
     """
-    return open("key.key", "rb").read()
+    return open(loc+"key.key", "rb").read()
 
 
-def encrypt(filename, key):
+def encrypt(loc, filename, key):
     """
     Given a filename (str) and key (bytes), it encrypts the file and write it
     """
     f = Fernet(key)
-    with open(filename, "rb") as file:
+    with open(loc+filename, "rb") as file:
         # read all file data
         file_data = file.read()
     # encrypt data
     encrypted_data = f.encrypt(file_data)
     # write the encrypted file
-    with open(filename, "wb") as file:
+    with open(loc+filename, "wb") as file:
         file.write(encrypted_data)
 
 
-def decrypt(filename, key):
+def decrypt(loc, filename, key):
     """
     Given a filename (str) and key (bytes), it decrypts the file and write it
     """
     f = Fernet(key)
-    with open(filename, "rb") as file:
+    with open(loc+filename, "rb") as file:
         # read the encrypted data
         encrypted_data = file.read()
     # decrypt data
     decrypted_data = f.decrypt(encrypted_data)
     # write the original file
-    with open(filename, "wb") as file:
+    with open(loc+filename, "wb") as file:
         file.write(decrypted_data)
